Finished Added Some Features

+++ app/src/main/java/com/kyawzinlinn/moviesapp/data/remote/dto/MovieDetailsDto.kt
@@ -17,13 +17,13 @@ data class MovieDetailsDto(
     val production_companies: List<ProductionCompany>,
     val production_countries: List<ProductionCountry>,
     val release_date: String,
-    val revenue: Int,
+    val revenue: Long,
     val runtime: Int,
     val spoken_languages: List<SpokenLanguage>,
     val status: String,
     val tagline: String,
     val title: String,
     val video: Boolean,
-    val vote_average: Double,
+    val vote_average: Float,
     val vote_count: Int
 )
